wip on hsv sliders; abstract slider & refactor
