+++ src/components/panel/GuiPanel.tsx
@@ -40,7 +40,7 @@ import ViewOptionsContext, {
 
 import useSupportedFilePicker from '@/hooks/useSupportedFilePicker';
 import { useModelSelectionExport } from '@/hooks';
-import GuiPanelTexture from './GuiPanelTexture';
+import GuiPanelTexture from './textures/GuiPanelTexture';
 import useSceneOBJFileDownloader from '@/hooks/useSceneOBJDownloader';
 import { mdiAxisArrow, mdiCursorDefaultOutline } from '@mdi/js';
 import Icon from '@mdi/react';
