+++ src/components/panel/textures/GuiPanelTexture.tsx
