+++ src/components/panel/textures/GuiPanelTextureMenu.tsx
@@ -4,9 +4,17 @@ import Menu from '@mui/material/Menu';
 import MenuItem from '@mui/material/MenuItem';
 import Icon from '@mdi/react';
 import { mdiDotsVertical } from '@mdi/js';
-import { styled } from '@mui/material';
+import {
+  Divider,
+  List,
+  ListItem,
+  Slider,
+  Typography,
+  styled
+} from '@mui/material';
 import { useFilePicker } from 'use-file-picker';
 import { replaceTextureDataUrl, useAppDispatch } from '@/store';
+import GuiPanelMenuSlider from '../GuiPanelMenuSlider';
 
 const StyledPanelTextureMenu = styled('div')(
   ({ theme }) => `& {
@@ -106,6 +114,30 @@ export default function GuiPanelTextureMenu({
             {option.label}
           </MenuItem>
         ))}
+        <Divider />
+        <List dense className={'hsv-sliders'}>
+          <GuiPanelMenuSlider
+            label={'H'}
+            min={-180}
+            max={180}
+            value={0}
+            onChange={() => undefined}
+          />
+          <GuiPanelMenuSlider
+            label={'S'}
+            min={-180}
+            max={180}
+            value={0}
+            onChange={() => undefined}
+          />
+          <GuiPanelMenuSlider
+            label={'L'}
+            min={-100}
+            max={100}
+            value={0}
+            onChange={() => undefined}
+          />
+        </List>
       </Menu>
     </StyledPanelTextureMenu>
   );
